Delete AbstractSquidEntity and move all its content into RocketSquidEntity
